ai-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
# LLM
groq>=1.1.2
# NOTE: Ollama is run as a separate process — no pip package needed
# Embeddings (upgraded from all-MiniLM to BGE)
sentence-transformers>=2.3.1
# Vector DB
faiss-cpu==1.7.4
# Document Parsing
pypdf==4.0.1
python-docx==1.1.0
# Web
python-multipart==0.0.6
aiohttp>=3.9.0
requests>=2.31.0
beautifulsoup4>=4.14.0
playwright>=1.41.0
# YouTube
youtube-transcript-api>=1.0.0
# Misc
python-dotenv>=1.0.1
numpy<2.0.0
